+++ cypress/component/rich-text/RichTextEditor.spec.ts
@@ -15,7 +15,7 @@ import { mountRichTextEditor } from './utils';
 
 // the sticky toolbar gets in the way of some of the tests, therefore
 // we increase the viewport height to fit the whole page on the screen
-describe('Rich Text Editor', { viewportHeight: 2000 }, () => {
+describe('Rich Text Editor', { viewportHeight: 2000, viewportWidth: 1000 }, () => {
   let richText: RichTextPage;
   let sdk: FieldAppSDK;
 
@@ -96,7 +96,7 @@ describe('Rich Text Editor', { viewportHeight: 2000 }, () => {
   });
 
   describe('history', () => {
-    it('supports undo and redo', () => {
+    it('supports undo and redo with keyboard shortcuts', () => {
       const expectedValue = doc(block(BLOCKS.PARAGRAPH, {}, text('some text.')));
 
       // type
@@ -110,6 +110,23 @@ describe('Rich Text Editor', { viewportHeight: 2000 }, () => {
 
       // redo
       richText.editor.click().type(`{${mod}}{shift}z`).click();
+      richText.expectValue(expectedValue);
+    });
+
+    it('supports undo and redo with toolbar buttons', () => {
+      const expectedValue = doc(block(BLOCKS.PARAGRAPH, {}, text('some text.')));
+
+      // type
+      richText.editor.click().type('some text.').click();
+
+      richText.expectValue(expectedValue);
+
+      // undo
+      richText.toolbar.undo.click();
+      richText.expectValue(undefined);
+
+      // redo
+      richText.toolbar.redo.click();
       richText.expectValue(expectedValue);
     });
 
@@ -220,17 +237,13 @@ describe('Rich Text Editor', { viewportHeight: 2000 }, () => {
 
       it('should add a new line after entity block in same list item', () => {
         richText.editor.click();
-
         richText.toolbar.ul.click();
 
-        richText.editor
-          .type('some text 1')
-          .type('{enter}')
-          .type(`{${mod}+shift+e}`)
-          .type('{enter}')
-          .type('some more text')
-          .type(`{${mod}+shift+e}`)
-          .type('{enter}');
+        richText.editor.type('some text 1').type('{enter}').type(`{${mod}+shift+e}}`);
+        richText.forms.embed.confirm();
+        richText.editor.type('{enter}').type('some more text').type(`{${mod}+shift+e}}`);
+        richText.forms.embed.confirm();
+        richText.editor.type('{enter}');
 
         richText.expectValue(newLineEntityBlockListItem);
       });