+++ packages/rich-text/src/plugins/shared/LinkedBlockWrapper.tsx
@@ -24,7 +24,7 @@ const styles = {
 type LinkedBlockWrapperProps = React.PropsWithChildren<{
   attributes: Pick<RenderElementProps, 'attributes'>;
   card: JSX.Element;
-  link: ResourceLink | EntityLink;
+  link: ResourceLink<'Contentful:Entry'> | EntityLink;
 }>;
 
 export function LinkedBlockWrapper({ attributes, card, children, link }: LinkedBlockWrapperProps) {
@@ -36,14 +36,12 @@ export function LinkedBlockWrapper({ attributes, card, children, link }: LinkedB
       data-entity-id={getLinkEntityId(link)}
       // COMPAT: This makes copy & paste work for Firefox
       contentEditable={IS_CHROME ? undefined : false}
-      draggable={IS_CHROME ? true : undefined}
-    >
+      draggable={IS_CHROME ? true : undefined}>
       <div
         // COMPAT: This makes copy & paste work for Chromium/Blink browsers and Safari
         contentEditable={IS_CHROME ? false : undefined}
         draggable={IS_CHROME ? true : undefined}
-        className={styles.container}
-      >
+        className={styles.container}>
         {card}
       </div>
       {children}